updated productForm,  moved shipping into product-detail
