+++ frontend/src/app/components/productForm/productForm.component.html
@@ -64,15 +64,17 @@
                 cdkTextareaAutosize
                 #autosize="cdkTextareaAutosize"
                 cdkAutosizeMinRows="1"
-                cdkAutosizeMaxRows="2"></textarea>
+                cdkAutosizeMaxRows="20"></textarea>
     </mat-form-field>
   </mat-card>
 
   <mat-card>
 
     <mat-form-field>
-      <input matInput placeholder="Product Price" [(ngModel)]="productPrice" name="productPrice">
-
+      <mat-label>Price</mat-label>
+      <input matInput placeholder="" [(ngModel)]="productPrice" name="productPrice">
+      <span matPrefix>CHF&nbsp;</span>
+      <span matSuffix>.00</span>
       <div *ngIf="empty(productPrice)" class="invalid-feedback">Please add a Price</div>
     </mat-form-field>
